fix: bug with additional commits message not parsed
This patch is a fix bug with additional Commits where the commit message
were not being passed to the parser correctly, this then caused an error
when trying to parse the commit message.

This patch also add a new output to the action called `versionType` which
will return the type of version that was bumped. This will be useful for
other actions to use.

also includes debug logs for the action, these were useful in debug our issue.